requirements.txt
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
Pillow==9.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pep8-naming==0.13.2
